Legger login-stuff for seg selv
